Enhance touchscreen script with error handling
Add checks for directory existence and improve file permissions.

+++ recovery/root/sbin/touchscreen.sh
@@ -1,8 +1,10 @@
 #!/sbin/sh
 
 # Read/Write mount default partitions
+    mkdir -p /system_root/system/media 2>/dev/null;
     mkdir -p /cache/touch 2>/dev/null;
     mkdir -p /data 2>/dev/null;
+    mount -o rw,remount /system_root || mount /system_root 2>/dev/null;
     mount -o rw,remount /vendor || mount /vendor 2>/dev/null;
     mount -o rw,remount /cache || mount -o rw /cache 2>/dev/null;
     mount -t ext4 /dev/block/platform/mtk-msdc.0/11230000.MSDC0/by-name/userdata /data || mount /data 2>/dev/null;
@@ -32,14 +34,12 @@ fi;
     while [ ! -d /data/media/0 ]; do 
       sleep 1
     done;
-    
-    mkdir -p /system_root/system/media
-
-    if [ -e /data/media/0/TWRP/bootanimation.zip ]; then
+    if [ -e /data/media/0/TWRP/bootanimation.zip ] && [ -d /system_root/system/media/ ]; then
       cp -vf /data/media/0/TWRP/bootanimation.zip /system_root/system/media/ && \
-      chmod 0775 /system_root/system/media/bootanimation.zip && \
+      chmod -v 0775 /system_root/system/media/bootanimation.zip && \
       chown 0:0 /system_root/system/media/bootanimation.zip && \
       chcon u:object_r:system_file:s0 /system_root/system/media/bootanimation.zip;
+      umount -l /system_root 2>/dev/null;
     fi;
     cp -f /cache/touch/recovery_touch_calibration /data/media/0/ && \
     chown 1023:1023 /data/media/0/recovery_touch_calibration && \
